+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.72](https://github.com/react4j/react4j/tree/v0.72) (2018-06-04)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.71...v0.72)